Use of combine data task publisher instead of data task completion closure!

+++ Dog Images/ImageStorage.swift
@@ -12,6 +12,19 @@ import UIKit
 struct ImageStorage {
     
     static let filemanager = FileManager.default
+    static private let imageFolderName = "Images"
+    
+    static var imageFolderUrl: URL? {
+        do {
+            let fileUrl = try filemanager.url(for: .documentDirectory, in: .userDomainMask, appropriateFor: nil, create: false).appendingPathComponent(imageFolderName)
+            return fileUrl
+            
+        }
+        catch {
+            print(error.localizedDescription)
+            return nil
+        }
+    }
       
     static func add(_ dogImage: DogImage) {
         
@@ -20,14 +33,15 @@ struct ImageStorage {
     
     static private func save(_ image: UIImage, with name: String) {
         
+        guard let folderUrl = imageFolderUrl else { return }
+        
+        let fileUrl = folderUrl.appendingPathComponent(name)
+        
+        guard let data = image.jpegData(compressionQuality: 0.8) else { return }
+        
         do {
-                        
-            let fileUrl = try filemanager.url(for: .documentDirectory, in: .userDomainMask, appropriateFor: nil, create: false).appendingPathComponent(name)
-                                    
-            guard let data = image.jpegData(compressionQuality: 0.8) else { return }
             
             try data.write(to: fileUrl)
-            
         }
         catch {
             
@@ -37,18 +51,18 @@ struct ImageStorage {
     
     static func getImages() -> [DogImage]?  {
         
+        guard let folderUrl = imageFolderUrl else { return nil }
+        
+        var dogImages = [DogImage]()
+        
         do {
-            
-            let fileUrl = try filemanager.url(for: .documentDirectory, in: .userDomainMask, appropriateFor: nil, create: false)
-            
-            var dogImages = [DogImage]()
-            
-            let contents = try filemanager.contentsOfDirectory(atPath: fileUrl.path)
+            let contents = try filemanager.contentsOfDirectory(atPath: folderUrl.path)
             
             contents.forEach { file in
                 
                 do {
-                    let data = try Data(contentsOf: fileUrl.appendingPathComponent(file))
+                    
+                    let data = try Data(contentsOf: folderUrl.appendingPathComponent(file))
                     guard let image = UIImage(data: data) else { return }
                     dogImages.append(DogImage(id: file, image: image))
                     
@@ -59,20 +73,24 @@ struct ImageStorage {
             }
             
             return dogImages
-            
         }
         catch {
             print(error.localizedDescription)
             return nil
         }
+    
     }
     
     static func removeImage(imageName: String) {
         
+        guard let folderUrl = imageFolderUrl else { return }
+        
+        let fileUrl = folderUrl.appendingPathComponent(imageName)
+        
         do {
-            let fileUrl = try filemanager.url(for: .documentDirectory, in: .userDomainMask, appropriateFor: nil, create: false).appendingPathComponent(imageName)
             try filemanager.removeItem(at: fileUrl)
         }
+        
         catch {
             print(error.localizedDescription)
         }
